Completely refactored game loop.
Much simpler, better tests, cleaner tests, and a defrecord for the
GameState. The big changes are a a defrecord for GameState, which has
state, transitions, and a quit flag. Transitions are side-effect
functions that can be returned by an update function. These are then
called by the main game loop. They don't update state directly - they
can fire events though.