Add the "laser"
In order to add the laser neatly I had to make a lot of changes. Mostly
around how I store the invaders, yet again. Now the invaders are a list
with offsets (rather than a row-column matrix). I initialize it from a
row column matrix, but otherwise it's just a bunch of hashes.

Now to make it move....